% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intcure.r
\docType{data}
\name{bmt}
\alias{bmt}
\title{Data from bone marrow transplant patients}
\format{
A data.frame object with columns:
\describe{
 \item{\code{time}:}{Disease Free Survival Time (Time To Relapse, Death Or End Of Study).}
 \item{\code{cens}:}{Relapse Indicator with 1-Relapsed, 0-Disease Free.}
 \item{\code{g}:}{Disease Group with 1-ALL, 2-AML Low Risk, 3-AML High Risk.}
 \item{\code{FAB}:}{1-FAB Grade 4 Or 5 and AML, 0-Otherwise.}
 \item{\code{inst}:}{Hospital with 1-The Ohio State University, 2-Alferd , 3-St. Vincent, 4-Hahnemann.}
}
}
\source{
Klein, J. P. & Moeschberger, M. L. Survival Analysis, Techniques for Censored and Truncated Data Spinger-Verlag, 2003
}
\usage{
data(bmt)
}
\description{
This data set is used in the examples of this package.
}
\examples{
data(bmt)
}
\keyword{datasets}
