% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intcure.r
\docType{data}
\name{bmtfit}
\alias{bmtfit}
\title{Model fit from a mixture cure model with random effects to bmt data}
\format{
A list of class \code{intcure}
}
\usage{
data(bmtfit)
}
\description{
This is the result of the model fit \code{intcure} as shown below
}
\examples{
# bmtfit = intcure(Surv(time, cens) ~ g + FAB + cluster(inst), ~ g + FAB,
# data = bmt, basedist = "weibull", sigma = c(-1, NA, NA),
# optimcfg = list(method = "BFGS", hessian = TRUE, maxit = 2000)).
data(bmtfit)
}
\keyword{datasets}
