% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intcure.r
\name{coef.intcure}
\alias{coef.intcure}
\title{Retrieve coefficients from a fitted mixture cure model with random effects}
\usage{
\method{coef}{intcure}(object, ...)
}
\arguments{
\item{object}{an \code{intcure} object obtained from \code{intcure}}

\item{...}{inherited but not used here}
}
\value{
a vector consists of \code{bt}, \code{gm}, \code{basepara} and \code{sigma}
}
\description{
It collects all the estimates from a fitted mixture cure model with random effects and forms a long vector.
}
\examples{
data(bmtfit)
coef(bmtfit)

}
\seealso{
\code{\link{intcure}}
}
\author{
Yingwei Peng
}
