% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intcure-package.R
\docType{package}
\name{intcure-package}
\alias{intcure-package}
\title{Random effects mixture cure models}
\description{
Fit random effects mixture cure models
}
\details{
This package fits a mixture cure model with random effects in both latency and incidence parts of the model. The random effects in the two parts can be correlated.
}
\references{
Peng, Y. and Taylor, J. M. G. Mixture cure model with random effects for the analysis of a multi-centre tonsil cancer study. Statistics in Medicine, 30:211-223, 2011
}
\author{
Yingwei Peng
}
