% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intcure.r
\name{print.intcure}
\alias{print.intcure}
\title{Print method for fitted mixture cure models with random effects}
\usage{
\method{print}{intcure}(x, ...)
}
\arguments{
\item{x}{an \code{intcure} object obtained from \code{intcure}}

\item{...}{other standard printing arguments}
}
\value{
It will return the object in \code{x}. This function is called for
side effects
}
\description{
It prints the estimates of the parameters in the model and the maximum likelihood value.
}
\examples{
data(bmtfit)
print(bmtfit)
}
\seealso{
\code{\link{intcure}}
}
\author{
Yingwei Peng
}
