% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intcure.r
\name{print.summary.intcure}
\alias{print.summary.intcure}
\title{Print method for summary of fitted mixture cure models with random effects}
\usage{
\method{print}{summary.intcure}(x, ...)
}
\arguments{
\item{x}{an \code{summary.intcure} object obtained from \code{summary()}}

\item{...}{other standard printing arguments}
}
\value{
It will return the object in \code{x}. This function is called for
side effects
}
\description{
It prints the estimates of the parameters, standard errors, z-values, p-values,
and the maximum likelihood value from a fitted mixture cure model with random effects.
}
\examples{
data(bmtfit)
print(summary(bmtfit))
}
\seealso{
\code{\link{intcure}}
}
\author{
Yingwei Peng
}
