% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intcure.r
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated Data}
\format{
A data.frame object with columns:
\describe{
 \item{\code{time}:}{Survival time}
 \item{\code{cens}:}{Censoring indicator}
 \item{\code{rx}:}{A binary treatment variable}
 \item{\code{id}:}{Cluster indicator}
}
}
\usage{
data(simdata)
}
\description{
This is a simulated data set for a testing purpose
}
\examples{
data(simdata)
}
\keyword{datasets}
